Feature: Add seed curriculum button to admin page (#2992)
Because:
* It's a more convenient way to test new lesson content than doing it
manually via the console
