+++ app/controllers/github_webhooks_controller.rb
@@ -6,7 +6,7 @@ class GithubWebhooksController < ApplicationController
   def github_push(payload)
     event = ::GithubPushEventAdaptor.new(payload)
 
-    UpdateLessonContentJob.perform_async(event.modified_urls) if event.merged_to_main?
+    Lessons::UpdateContentJob.perform_async(event.modified_urls) if event.merged_to_main?
   end
 
   private
