+++ app/jobs/lessons/import_all_content_job.rb
@@ -1,0 +1,10 @@
+module Lessons
+  class ImportAllContentJob
+    include Sidekiq::Worker
+    sidekiq_options retry: 1, dead: false
+
+    def perform
+      LessonContentImporter.import_all
+    end
+  end
+end
